Check it is a split jump from one full square to two empty squares
